requirement.txt
flask
flask_sqlalchemy
flask-cors
psycogy-2
python-dotenv
Flask-Bcrypt